refactor : 확장자 등록 / 삭제 시 유효성 검사 추가
- 확장자 길이 제한 명시적 유효성 검사처리
- 삭제시 Default 확장자 삭제 불가처리

+++ src/main/java/assignment/flow/application/BlockExtensionCommandServiceImpl.java
@@ -71,6 +71,10 @@ public class BlockExtensionCommandServiceImpl implements BlockExtensionCommandSe
             throw new BlockExtensionLimitException();
         }
 
+        if (name.isEmpty() || name.length() > 20) {
+            throw new IllegalArgumentException();
+        }
+
 
         BlockExtension blockExtension = new BlockExtension(name, ExtensionType.CUSTOM, true);
 
@@ -83,6 +87,10 @@ public class BlockExtensionCommandServiceImpl implements BlockExtensionCommandSe
     public void deleteExtension(String extensionName) {
         BlockExtension blockExtension = repository.findByExtensionName(extensionName)
                 .orElseThrow(BlockExtensionNotSubject::new);
+
+        if (blockExtension.getExtensionType() == ExtensionType.DEFAULT) {
+            throw new IllegalArgumentException();
+        }
 
         repository.delete(blockExtension);
     }
